backend/requirements.txt
# Core backend
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlmodel>=0.0.27
sqlalchemy>=2.0.40
psycopg2-binary>=2.9.10

# Environment & settings
python-dotenv>=1.1.0
pydantic[email]>=2.5.0
pydantic-settings>=2.1.0
PyYAML>=6.0.2

# Security
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0

# File uploads
python-multipart>=0.0.18

# Requests
requests>=2.32.3

# LangChain ecosystem (optional if using LLMs)
langchain>=0.3.49
langchain-openai>=0.3.5
langchain-core>=0.3.22
openai>=1.76.0
tiktoken>=0.9.0

# Gemini compatibility
google-generativeai>=0.8.0
